Refactor PMApplication and update service layers
This commit includes significant refactoring of the PMApplication and related projects, updating project references, namespaces, and class structures. Multiple new DTOs have been added, such as `ExportSkuDto`, `JobDto`, and `StandTypeDto`, to enhance data handling.

Services like `PlanogramService`, `OrderWindowService`, and `RegionService` have been updated with new methods and filters for improved functionality. User authentication is now managed using `ClaimsPrincipal`, replacing previous methods for user information retrieval.

Interfaces `IOrderWindowService` and `IRegionService` have been modified to include asynchronous methods, enhancing performance. The `PlanMatrContext` class has been updated to manage new entities, and various controllers have been refactored to utilize the new DTOs and services.

Overall, these changes improve code organization, maintainability, and support better separation of concerns for future development.

+++ Entities/OrderAggregate/OrderItemInfo.cs
@@ -21,6 +21,6 @@ namespace PMApplication.Entities.OrderAggregate
         public bool Shoppable { get; set; }
         public bool HidePrices { get; set; }
         public bool? IsFullPlano { get; set; }
-        public int? OrderPlanogramId { get; set; }
+        public long? OrderPlanogramId { get; set; }
     }
 }
